third_party/WebKit/LayoutTests/fast/dom/setAttributeNS-prefix-and-null-namespace-expected.txt

Test that calling setAttributeNS() with a prefixed qualifiedName and null NS throws NAMESPACE_ERR.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElement('test').setAttributeNS(null, 'foo:bar', 'baz') threw exception NamespaceError: An attempt was made to create or change an object in a way which is incorrect with regard to namespaces..
PASS successfullyParsed is true

TEST COMPLETE
